Separate copy button into own file.
