+++ site/index.html
@@ -31,7 +31,8 @@
   <body>
     <div id="app"></div>
     <script src="/static/js/config.js"></script>
-    <script src="/static/js/plugins.js"></script>
+    <script src="/static/js/plugins/plugins.js"></script>
+    <script src="/static/js/plugins/copy_buttons.js"></script>
     <script src="//cdn.jsdelivr.net/npm/docsify/lib/docsify.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/docsify-themeable@0/dist/js/docsify-themeable.min.js"></script>
     <script src="//cdn.jsdelivr.net/npm/docsify-pagination/dist/docsify-pagination.min.js"></script>
